tools/utils: Move string-to-int helper into a separate file
