+++ tools/bl.c
@@ -16,7 +16,6 @@
 
 #include <inttypes.h>
 #include <stdbool.h>
-#include <assert.h>
 #include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
@@ -30,46 +29,9 @@
 #include "device.h"
 #include "msg.h"
 #include "sig.h"
+#include "util.h"
 
 typedef int (* bl_cmd_fn)(int argc, char *argv[]);
-
-static inline bool check_fit(uint32_t value, size_t target_size)
-{
-	uint32_t target_max;
-
-	assert(target_size <= 4);
-
-	target_max = (~(uint32_t)0) >> ((sizeof(uint32_t) - target_size) * 8);
-
-	if (value > target_max) {
-		return false;
-	}
-
-	return true;
-}
-
-static inline bool read_sized_uint(
-		const char *value,
-		uint32_t *out,
-		size_t target_size)
-{
-	unsigned long long temp;
-	char *end = NULL;
-
-	errno = 0;
-	temp = strtoull(value, &end, 0);
-
-	if (end == value || errno == ERANGE || temp > UINT32_MAX) {
-		return false;
-	}
-
-	if (!check_fit(temp, target_size)) {
-		return false;
-	}
-
-	*out = (uint32_t)temp;
-	return true;
-}
 
 int bl_cmd_read_and_print_message(int dev_fd, int timeout_ms)
 {
